requirements.txt
pinecone
transformers
torch
Pillow
requests